requirements.txt
selenium==4.15.2
undetected-chromedriver==3.5.4
beautifulsoup4==4.12.2
pandas==2.1.3
python-dotenv==1.0.0
supabase==2.4.0
webdriver-manager==4.0.1
httpx>=0.20.0 